Faas 6d: tõsta public access ja SEO routerisse #36
